filter on distance, not time
Filter sequential entries which are closer than 5 meters to each
other. This seems to be a better policy than separating entries by
time, the only setback is that if someone takes pictures constantly
moving such that distance between pictures is less than 5 meters,
those will be filtered out.

Not a big deal.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>